pyproject.toml
[tool.poetry]
name = "contractshark"
version = "0.2.0"
description = "smart contract monitoring"
authors = ["sam <sam@freighttrust.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
"Django-2.2.19" = "2.0.1"
djangorestframework = "3.7.7"
celery = "4.1.0"
spotlight-sqlalchemy = "^1.0.2"
mythril = "0.10.6"
gunicorn = "19.7.1"
dj-database-url = "0.4.2"
psycopg2 = "2.7.3.2"
django-celery-results = "1.0.1"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
